requirements.txt
fastapi
uvicorn
mygeotab
python-dotenv
pandas
duckdb
googlemaps
pydantic_ai
